+++ core/src/main/java/pfg/kraken/ConfigInfoKraken.java
@@ -48,6 +48,7 @@ public enum ConfigInfoKraken implements ConfigInfo
 	MAX_LATERAL_ACCELERATION(3), // maximal lateral acceleration, in m/s²
 	MAX_LINEAR_ACCELERATION(2), // maximal linear acceleration, in m/s²
 	DEFAULT_MAX_SPEED(1), // in m/s (or mm/ms)
+	MINIMAL_SPEED(0), // in m/s
 	MAX_CURVATURE(5), // in m⁻¹
 	STOP_DURATION(800), // temps qu'il faut au robot pour s'arrêter et repartir
 						// (par exemple à cause d'un rebroussement) in ms