+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # Unreleased
 
+- add in configuration the MINIMAL_SPEED key. The maximal curvature is limited in order to generate paths that satisfy this minimal speed.